requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
asgiref==3.10.0
Authlib==1.6.5
bandit==1.8.6
billiard==4.2.2
binaryornot==0.4.4
black==25.11.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cookiecutter==2.6.0
cron_descriptor==2.0.6
cryptography==46.0.3
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.8
django-axes==8.0.0
django-browser-reload==1.21.0
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-cotton==2.3.1
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.2
django-htmx==1.26.0
django-import-export==4.3.13
django-redis==6.0.0
django-tailwind==4.4.1
django-timezone-field==7.1
djangorestframework==3.16.1
dparse==0.6.4
filelock==3.20.0
flake8==7.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.2
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nltk==3.9.2
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytailwindcss==0.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytokens==0.3.0
PyYAML==6.0.3
redis==7.0.1
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stevedore==5.5.0
tablib==3.9.0
tenacity==9.1.2
text-unidecode==1.3
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
crispy-bootstrap5==2025.6
